requirements.txt
pandas
flask
scikit-learn
numpy 
scipy
requests
minio
tensorflow
matplotlib
statsmodels
pyyaml

